Move builtin extensions
