+++ construct/ext/__init__.py
@@ -3,9 +3,6 @@
 from __future__ import absolute_import
 
 # Local imports
-from ..hou.host import Houdini
-from ..maya.host import Maya
-from ..nuke.host import Nuke
 from .cache import Cache, UserCache
 from .software import Software
 
@@ -14,7 +11,4 @@ extensions = [
     Software,
     Cache,
     UserCache,
-    Maya,
-    Houdini,
-    Nuke
 ]
