+++ construct/hosts/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import
+
+# Local imports
+from .hou.host import Houdini
+from .maya.host import Maya
+from .nuke.host import Nuke
+
+
+extensions = [
+    Houdini,
+    Maya,
+    Nuke,
+]
