+++ construct/hosts/maya/icons/maya.png
