+++ construct/hosts/nuke/icons/nuke.png
