+++ construct/hosts/nuke/software/nuke.yaml
