+++ construct/hosts/nuke/software/nukex.yaml
